Workaround local development environment weirdness + fix config.h.in
Something with macOS Ventura 13.3 and Xcode 14.3 breaks init_hookFunctions(): the app compiles fine but won't launch, throwing off errors that were supposedly fixed already. This variable allows me to disable the function outright so I can continue UI development without having to downgrade to Ventura 13.2.1 (because Time Machine is going to be an ASS to deal with.) Don't set this variable as it will cause the game itself to not behave correctly.