Feat[SurfaceViewController] Add two-finger long press to show and hide keyboard
Since the entire GUI is set to automatically hide when a controller is connected, this should allow users to be able to show the keyboard without needing to rebind keys. Can also lead to removal of the keyboard button itself since that is no longer needed. More gesture recognizers for other buttons (Debug, Fullscreen, etc) may come in the future.